configdict: T2660: use Config to prevent duplicate
The ConfigDict matches fully the Config API for:
get_level, set_level and _make_path.

The code was re-implementing these function but with set_level
spliting the path using split() and not re (which has a slightly
different semantic and output).

Converted the calls to be passthough so make sure that the
behaviour remain consistent between both classes.