mangling: T2660: Config store mangling logic
The mangling was previously performed when calling function
to extract the dict. Therefore it was possible to use Config
with mangling but get un-modified result when using ConfigDict

This problem also existed for the XML code which could not
figure out if it should modify the keys or not when used.

This patches adds a function to Config called mangler which
is then used when set to update the dictionary when it is requested.

ConfigDict code can then use to making sure key names will be
consistent.